Fix course users remove type migration
This is a fix from an earlier half-baked attempt at multiple instructors

+++ db/migrate/20160126233939_remove_type_from_course_user.rb
@@ -1,5 +1,7 @@
 class RemoveTypeFromCourseUser < ActiveRecord::Migration
   def change
-    remove_column :course_users, :type
+    if column_exists? :course_users, :type
+      remove_column :course_users, :type
+    end
   end
 end
